refactor: organize Cart Service and Controller in a cart paste
